src/services/__tests__/affiliateNetworkService.test.ts
import { affiliateNetworkService } from '../affiliateNetworkService';
import { UnifiedIntelligenceField } from '../../utils/unifiedIntelligenceField';

jest.mock('../../utils/unifiedIntelligenceField');

describe('AffiliateNetworkService', () => {
  beforeEach(() => {
    jest.clearAllMocks();
  });

  describe('signUpForNetwork', () => {
    it('should successfully sign up for a network', async () => {
      const mockUIF = UnifiedIntelligenceField.getInstance as jest.Mock;
      mockUIF.mockReturnValue({
        performEthicalCheck: jest.fn().mockResolvedValue(true),
        encryptSensitiveData: jest.fn().mockResolvedValue('encrypted'),
        generateSecureKey: jest.fn().mockResolvedValue('key'),
        determineOptimalSignupStrategy: jest.fn().mockResolvedValue('strategy'),
        executeNetworkSignup: jest.fn().mockResolvedValue({ success: true, credentials: {} }),
        updateConnectivityLevel: jest.fn()
      });

      const result = await affiliateNetworkService.signUpForNetwork('TestNetwork', { name: 'Test User' });

      expect(result).toEqual({ success: true, message: 'Successfully signed up for TestNetwork' });
    });

    it('should handle ethical check failure', async () => {
      const mockUIF = UnifiedIntelligenceField.getInstance as jest.Mock;
      mockUIF.mockReturnValue({
        performEthicalCheck: jest.fn().mockResolvedValue(false)
      });

      const result = await affiliateNetworkService.signUpForNetwork('TestNetwork', { name: 'Test User' });

      expect(result).toEqual({ success: false, message: 'Signup process did not pass ethical check' });
    });

    it('should handle signup failure', async () => {
      const mockUIF = UnifiedIntelligenceField.getInstance as jest.Mock;
      mockUIF.mockReturnValue({
        performEthicalCheck: jest.fn().mockResolvedValue(true),
        encryptSensitiveData: jest.fn().mockResolvedValue('encrypted'),
        generateSecureKey: jest.fn().mockResolvedValue('key'),
        determineOptimalSignupStrategy: jest.fn().mockResolvedValue('strategy'),
        executeNetworkSignup: jest.fn().mockResolvedValue({ success: false, error: 'Signup failed' }),
        updateConnectivityLevel: jest.fn()
      });

      const result = await affiliateNetworkService.signUpForNetwork('TestNetwork', { name: 'Test User' });

      expect(result).toEqual({ success: false, message: 'Signup failed' });
    });

    it('should handle unexpected errors', async () => {
      const mockUIF = UnifiedIntelligenceField.getInstance as jest.Mock;
      mockUIF.mockReturnValue({
        performEthicalCheck: jest.fn().mockRejectedValue(new Error('Unexpected error'))
      });

      const result = await affiliateNetworkService.signUpForNetwork('TestNetwork', { name: 'Test User' });

      expect(result).toEqual({ success: false, message: 'An unexpected error occurred during signup' });
    });
  });

  // Add more tests for other methods...
});